Fixing a slice out of bounds error
The error is caused by a bug in the compressor. When the match sequence
pointer starts 256 entries back into the buffer, its byte value when
written into the output file is not 256 but 0. This invalidates the
compressed sequence and causes the decompressor to fail.

I will not be investing more time on the LZ77 implementation thus the
quick fix is to reduce the capacity of the lookahead buffer and the
window from 256 to 255 bytes. I also increased the input and output
buffers.